requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
cryptography>=42.0.0
slowapi>=0.1.9
playwright==1.48.0
httpx>=0.27.0
python-jobspy>=1.1.0
celery[redis]==5.4.0
redis==5.1.0
thefuzz[speedup]==0.22.1
aiosqlite==0.20.0
stripe>=9.0.0

# Dev
ruff>=0.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
